Separate bot.py into smaller extensions.
